HHH-13069 Go back to non-SSL links for our JBoss Nexus repositories
Using SSL repositories with JDK 11 ends up with the error below:
javax.net.ssl.SSLHandshakeException: extension (10) should not be presented in server_hello

This is a bug in JDK 11 that will apparently be fixed in the next update
(planned for January 2019).

Until then, we need to use non-SSL links.

For more information:
https://stackoverflow.com/questions/52016415/jdk-11-ssl-error-on-valid-certificate-working-in-previous-versions
https://bugs.openjdk.java.net/browse/JDK-8209965 - JDK 12 issue
https://bugs.openjdk.java.net/browse/JDK-8210005 - JDK 11 issue